(MODULES-5370) Allow arguments to be passed in
Prior to this commit, the path to the module CHANGELOG and metadata
were hard coded because the task was intended to be run from within
a module directory.

Since the tool will now be used external to a module, add arguments
so that the path to the module can be passed in. Also allow the
latest tag to be passed in as an argument rather than an environment
variable.